Fix a bug that caused external games to appear active after the session ends.
