+++ src/Gameboard.Api/Features/Game/Requests/GetGamePlayState/GetGamePlayState.cs
@@ -1,11 +1,15 @@
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Gameboard.Api.Common.Services;
+using Gameboard.Api.Data;
 using Gameboard.Api.Features.Teams;
 using Gameboard.Api.Services;
 using Gameboard.Api.Structure.MediatR;
 using Gameboard.Api.Structure.MediatR.Authorizers;
 using Gameboard.Api.Structure.MediatR.Validators;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 
 namespace Gameboard.Api.Features.Games.Start;
 
@@ -16,6 +20,8 @@ internal class GetGamePlayStateHandler : IRequestHandler<GetGamePlayStateQuery, 
     private readonly EntityExistsValidator<GetGamePlayStateQuery, Data.Game> _gameExists;
     private readonly IGameModeServiceFactory _gameModeServiceFactory;
     private readonly IGameService _gameService;
+    private readonly INowService _now;
+    private readonly IStore _store;
     private readonly ITeamService _teamService;
     private readonly EntityExistsValidator<GetGamePlayStateQuery, Data.User> _userExists;
     private readonly UserRoleAuthorizer _userRoleAuthorizer;
@@ -26,6 +32,8 @@ internal class GetGamePlayStateHandler : IRequestHandler<GetGamePlayStateQuery, 
         EntityExistsValidator<GetGamePlayStateQuery, Data.Game> gameExists,
         IGameModeServiceFactory gameModeServiceFactory,
         IGameService gameService,
+        INowService now,
+        IStore store,
         ITeamService teamService,
         EntityExistsValidator<GetGamePlayStateQuery, Data.User> userExists,
         UserRoleAuthorizer userRoleAuthorizer,
@@ -35,6 +43,8 @@ internal class GetGamePlayStateHandler : IRequestHandler<GetGamePlayStateQuery, 
         _gameExists = gameExists;
         _gameModeServiceFactory = gameModeServiceFactory;
         _gameService = gameService;
+        _now = now;
+        _store = store;
         _teamService = teamService;
         _userExists = userExists;
         _userRoleAuthorizer = userRoleAuthorizer;
@@ -60,6 +70,31 @@ internal class GetGamePlayStateHandler : IRequestHandler<GetGamePlayStateQuery, 
         })
         .AddValidator(_userExists.UseProperty(r => r.ActingUserId));
         await _validatorService.Validate(request, cancellationToken);
+
+        // default rules that apply to all sessions
+        var teamSession = await _store
+            .WithNoTracking<Data.Player>()
+            .Where(p => p.TeamId == request.TeamId)
+            .Select(p => new
+            {
+                p.SessionBegin,
+                p.SessionEnd,
+                p.Role
+            })
+            .ToArrayAsync(cancellationToken);
+
+        var begin = teamSession.Select(p => p.SessionBegin).Distinct().Single();
+        var end = teamSession.Select(p => p.SessionEnd).Distinct().Single();
+
+        if (begin.IsEmpty())
+            return GamePlayState.NotStarted;
+
+        var nowish = _now.Get();
+        if (begin <= nowish && (end.IsEmpty() || end >= nowish))
+            return GamePlayState.Started;
+
+        if (nowish > end)
+            return GamePlayState.GameOver;
 
         var modeService = await _gameModeServiceFactory.Get(gameId);
         return await modeService.GetGamePlayStateForTeam(request.TeamId, cancellationToken);
